Added plant mode basic UI
