+++ My Island/Assets/_Main/Assets/Prefabs/UI/PlantMode.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8e8ddd9382da381449969fc6071184c8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
