+++ My Island/Assets/_Main/Assets/Prefabs/UI/PlantMode/PlantMode.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 528c7c30b5a1aea4cb7b0c55c62fb670
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
