+++ My Island/Assets/_Main/Scripts/Features/Forest.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 196d9c0e5dd848545a1c5141fa0a2324
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
