+++ My Island/Assets/_Main/Scripts/Features/Forest/ForestDataStructures.cs
@@ -1,0 +1,10 @@
+namespace MyIsland
+{
+    public class ForestData
+    {
+        public ForestData()
+        {
+            
+        }
+    }
+}
