+++ My Island/Assets/_Main/Scripts/Features/Forest/ForestSources.cs
@@ -1,0 +1,14 @@
+using System;
+using UniRx;
+
+namespace MyIsland
+{
+    public interface IForestSource
+    {
+        IObservable<bool> IsPlantModeOpen { get; }
+        ForestData Data { get; }
+
+        void EnterPlantMode();
+        void ExitPlantMode();
+    }
+}
