+++ My Island/Assets/_Main/Scripts/UI/PlantMode.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3f67d16099fb2df4c98c47458b35bd4d
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
