+++ My Island/Assets/_Main/Scripts/UI/PlantMode/PlantModeUIController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 8c6d180d4aef7d044917ab74e0e6c57c
