--- My Island/Assets/_Main/Scripts/UI/Splash/SplashScreenController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: b242b61f3fbda4d02afeca4bd48ae707
