+++ My Island/Assets/_Main/Scripts/UI/Splash/SplashScreenUIController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: b242b61f3fbda4d02afeca4bd48ae707
