--- My Island/Assets/_Main/Scripts/UI/TellMeYourName/TellMeYourNameController.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 17db6a89bb30ad6408da158a8649a6b4
