+++ My Island/Assets/_Main/Scripts/UI/TellMeYourName/TellMeYourNameUIController.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 17db6a89bb30ad6408da158a8649a6b4
